R4R:modify voting_power_for_height method

+++ backend/vo/proposal.go
@@ -88,7 +88,7 @@ type (
 		InitialDeposit   Coin                `json:"intial_deposit,omitempty"`
 		TotalDeposit     Coin                `json:"total_deposit,omitempty"`
 		Votes            []VoteWithVoterInfo `json:"votes"`
-		TotalVotingPower int64               `json:"voting_power_for_height"`
+		TotalVotingPower float64             `json:"voting_power_for_height"`
 		SubmitTime       time.Time           `json:"submit_time,omitempty"`
 		DepositEndTime   time.Time           `json:"deposit_end_time,omitempty"`
 		VotingEndTime    time.Time           `json:"voting_end_time,omitempty"`
